Added generating model code from templates
- Jinja2 templates can now be used for code generation.
- fixed a few issues with the code generation
- Created a new 'Forgather' notebook for testing syntax.
- Updated syntax documentation

Known Issues:
- lambda code generation does not work when the lambda takes arguments,
although it works correctly for materialization now. Resolving this
looks to be difficult.

- Loading models into the same process with different automatically
generated code can result in the process retaining the old model
constructor, as the import system can't tell it's not the same code.
Mitigation: Restart Python before training a new model.

- generated lambda code, which defines factory nodes used outside the
scope of the lamda don't resolve, as the definition is in the lambda.
This can be resolved by ensuring that shared factories are first used
outside the scope of the lambda. Not ideal, but it works for now.

- Renamed !callable to !factory, as it's a better description.