Partially fixed issues with lambdas
The main Latent.materialize() method was not forwarding variables to
lambdas within the enclosing scope. This was resolved by return a
partial function, to wrap the CallableNode, which forwards args from the
outer scope, while also allowing additional arguments to be passed to
the partial function. This should be much closer to the semantics of
lambdas, which allow access to variables in the enclosing scope.

The was an insue with the code generator, where reused definitions would
be declared at their first occurance in the graph, then the remaining
references would reference a temporary variable: if the first occurance
was inside the scope of a lambda, the definition would not be visible
outside of the lambda.

This was resolved by refactoring the 'main_body' code output into
'definitions' and 'main_body,' where all needed definitions are declared
in depth-first order, before the main_body.

The result is that these are all declared in dependency order.