Resolved issue with dynamic imports in code
Performing dynamic imports at the start of a generated file could cause
issues when the import depends upon recursively generated code having
been created.

Resolved by making the dynamically imported symbols lambdas, which are
only called when the symbol actually needs to be resolved.

There is still a remaining issue with automatically copying source files
when running a training script as code. It appears that the relative
imports of the generated code are not recognized as part of the model's
module, thus they are not copied to the model directory.

The reason likely has something to do with differences in how 'from
<module> import <symbol>' differs from how forgather.dynamic imports
these modules. TBD.