Added new 'workspace' function
The meta-config recursively searches directories, starting with the project
directory, for a directory named 'forgather_workspace,' which is then
added as a template searchpath for the meta-config.

This makes is much easier to have a common default meta-config for a
group of projects under the same 'workspace' directory, making it easier
to update common settings in one place.

The workspace directory root location is also injected configuration
templates, allowing them to compute relative directory locations, like
common workspace template directories.