Fixed batch typing and renamed tutorials
Datasets can yield either dictionaries or sequences. I had incorrectly
assumed that the sequences were tuples, when they were actually a list,
which resulted in the type being assumed to be a dictionary.

As to handle either tuples or lists, we now check for ABC Sequence.

It also looks like some of my updates which I though had been pushed
with the previous commit are still pending -- working on the tutorials,
so updating that too.