Reordered the arguments to Project()
A common use-case is to construct a project, where working directory is
the project directory. When required to specify the project directory
first, it's more awkward to just load a target configuration, as you
either need to use the keyword or explicilty specify the project path.

Reversing the order of these arguments makes this eaiser to work with.

Other changes:
- Fixed (hopefully) the same bug from the prior commit in AccelTrainer
- Renamed the tutorial directories
- Updated code which would break based upon the previous Project() API.
- Added support for Project to have a config of 'None' and added an
explicit 'load_config()' method, which can change the project config
without having to construct a new instance.