Added nproc hint and fixed training script
The basic trainer does not support multi-GPU configurations and will not
work right, if started with multiple processes. As it is, this requires
one to manually set the number of training processes to 1 when using
torchrun -- or set -n 0, when using the fgcli.

This change adds meta-data to the training script to hint at the correct
number of procs-per-node to use and fglic reads this value and sets in
when invoking torchrun.

To make the training script more robust, when invoked without torchrun,
it defaults to rank-0, if RANK is not set in the envrionment. This
allows it to be run without torchrun for the simple case of the basic
trainer class.