Fix RoPE, template cleanup, fix wall-of-text
The RoPE implementation, written by Claude, appeared to be working, but
upon closer inspection, the implementation had retained a seperate RoPE
embedding for every layer, rather than passing it up from the causal_lm
layer via forwarding.

I removed the code to insert RoPE layer in every attention module and
enabled the RoPE causal_lm module, which had not been tested. That
implementation had a couple of bugs, which I fixed. Now we only have a
single instance of the embeddings, which are passed via the forward()
methods through the entire stack. This seems to be working now.

I had noted earlier than there is something of a definciency in how the
Huggingface library handles dynamic imports for custom models. That is,
it only looks one level deep to determine which files should be
included. Claude's implementation was importing the rotary embedding
module indirectly, which caused the custom code to not work when loaded
with the HF interface. This can be addressed by using dependency
injection at the top level, thus all imports stay in the top-level file
and model loading does not fail.

While working on this and looking at the structure, it was apparent that
I could fully factor out the RoPE implementation and make the modules
work with any relative positional encoding implementation via dependency
injection.

Thus, the new attention module and the new RoPE causal model have been
made more general, have been renamed appropriately, and, in theory,
could make use of learnable relative embeddings, like the T5 model. This
could be a fun project to experiment with!

One issue remains with the rotary embeddings. When we try to save them
at the end of the training, the safetensors implementation barfs and
raises an exception. It looks like it's an issue with supporting complex
tensors. This requires further investigation. I suspect that switching
to saving in torch's native format will address this, but it's strange
than safetensors can't handle complex tensors!?

While running training sessions and having a wall-of-text dumped on each
error, I decided to track this down, as this has been a persistent
annoyance. Essentially, it was dumping the full text of the
configuration multiple time on each failure, which make debugging
difficult. I found the source of the problem and have eliminated the
wall-of-text issue.

While cleaning up the new configuraiton, I noticed that the
dynamic_causal_model template defintion had a number of singleton object
instances, used when constructing the model. The result is that when
multiple model instance are constructed via the model factory, they will
share some common torch modules, which is definitely a bad thing. This
had been corrected by changing them from singleton nodes to factory
nodes, which create a new instance each time they are referenced.

The above issue seems to have been created when I switched from passing
a single, constructed, model instance to the trainers to passing a model
factory -- something needed for the pipeline trainer. Apparenlty I
misssed the issue at the time.

While looking for any similar bugs in the templates, I identified quite
a few instances of the depricated "!lambda" tag, which I updated to
"!partial." I did not actually run all of the modified configs, so it's
possible that a typo could have broken something. We really should have
something which can at lease automatically reparse all of the example
for validity. One more thing for the TODO list...

While I was at it, I created a couple of new configurations for the
positional_encoding_comparison project: no-positional-encoder and Alibi.
I ran them and reproduced some my older findings that no-pe works
surprising well. Overall, Alibi did the best. It's worth running a much
longer training session, with the full dataset, over several epochs. I'm
curious how each does in the long run?

Next of the agenda is to replace /most/ of the attention implementations
with an injectable imlementation with the same interface as
torch.nn.functional.scaled_dot_product_attention(). See: https://docs.pytorch.org/docs/stable/generated/torch.nn.functional.scaled_dot_product_attention.html#torch.nn.functional.scaled_dot_product_attention