Fixed a few issues with the tiny_llama tutorial
Problem:
It appears the HF library has an issue with complex tensors, which
we use in our RoPE implementation. Safetensor can't save them and the HF
load_from_pretrained(), when using Pytorch file format, insists on
casting the complex tensors to real, discarding the imaginary
component. This breaks things.

The tutorial is also a bit out of date. We should really show off the
dynamic model implementation, as it's a major benefit of the framework.

Solution:
Save tensors using pytorch format. This only partially mitigates the
issue. To address the cast to real, we use the forgather load method to
load the weights from the checkpoint. This works, but is not portable.

TODO: Find a better long-term solution, even if that is with a PR to the
HF library.

The tutorial has been reorganized around around the dynamic
implementation, with the HF model being a secondary target now. Also
removed the larger version of the model, as this should be a quick demo.

To demonstrate the dynamic nature of the model, I have added a config
which replaces the pre-layer-norm implementation with post-layer-norm.