Updated Llama 7B finetune example
With the recent changes to checkpoint handling, the config for the Llama
7B finetuning exmaple needed to be updated.

While doing so, I modified the code for loctating the latest checkpoint
to default to looking in the model's main directory, if the
"checkpoints" directory does not exist. This simplifies the use case of
converting a model, then training it.

Tested the trainer to the first checkpoint, interrupted the process, and
resumed training from the checkpoint without issue.