Added Samantha dataset and updated pipeline llama
I added a dataset definition for Samantha. See https://erichartford.com/meet-samantha
This should be a bit more interesting when testing the pipeline trainer
with larger models.

Added more pipeline types to the llama pipe example.

Added support for chat-templates, which are needed for preprocessing
this type of dataset.

Added trainer module for tracking and dumping peak memory usage to help
tune parameters.

Updated llama pipe configuration to use Samantha and tuned setting a
little. There's probably more work to be done, but it is working.