Updated Activation Checkpointing Example/Test
To verify that activation checkpointing still works, I added a new large
model definition, the Deep One (a Deepnet model with ALiBi attetnion)
and updated the checkpointing example.

I also modified Tiny Experiments to make it easier to configure peak
memory debugging in derived projects.

And for fun, I added a "Tiny Deep One" model to Tiny Experiments and
added it to the Tiny Models project. It actually does better than any of
the other models!