Renamed fgcli.py to 'forgather'; as console-script
Updated the pyproject.toml to install the Forgather CLI as "forgather,"
and removed the janky "bin" directory and shim binary.

This required updated all of the instructions, which were updated with
find/sed.

It seems to work, both in my original base worksapce and in a venv, as a
fresh install.