Memory Budget and Fused Backward/Optimizer
Added support for the new PyTorch "memory_budget" API, which
selectively applies activation checkpointing, when using
torch.compile(), according to a specified "memory-budget" parameter.

I have also added the option to fuse the gradient computation with the
optimizer step, for additional memory savings.

As the pipeline trainer uses torch.export() to split the model, any
manually specified activation checkpoints will be stripped from the
model. One way of addressing this is to use the memory-budget API and
torch compile. I have also added an experimental function to modify the
fx graph by splicing in checkpoints at each layer, which provides
comparable functionality to the manual method.

One noted issues is that when using the fused backward/optimizer step,
the saving are not realized when not using explicit activation
checkpointing -- either manually added to the model or applied through
graph manipulation. My suspicion is that something in autograd is
holding references to the gradients until the full backward step
completes or until the operation is split up via explicit checkpoints.
The root cause is TBD -- and it's annoying!

Other changes:
- Cut down on much of the pipeline trainer debug messages at the "info"
level.
- Added support for explicitly controlling the SDPA implementation used
by PyTorch.
- Added support for setting set_float32_matmul_precision() torch option,
which improves 32-bit performance substantially on Ampere or higher
GPUs.
- Added option for enabling torch autograd anomaly detection, for
debugging.
- Removed all "activation_checkpoint" project and replaced it with a new
"peak_memory" project, where we compare various techniques for saving
memory, including the new features.
- Updated llama_pipeline to at least call-out the new options in the
project config. These have been manually tested, but configurations
still need to be created to demonstrate how to do it.
-Modified how CLI handles configuration paths, as to allow
sub-directories of the config. directory.
- Fixed issues in the peak_memory callback and updated projects that use
it.