Added support for CPU activation offloading
This redues GPU memory usage by stashing activation memory to the CPU
and restoring at the backward() step. This option should be combined
with some form of activation checkpointing, as doing this with every
activation is really slow.

In any case, when combined with other memory saving options, this should
allow for training with very long sequences -- if you have the time for
it!

Added new example configs with this feature, which also shows how to use
the memory saving features with the pipeline trainer.