Added new finetuning examples project
This is the start of a new example which is intended to demonstrate how
to use Forgather for finetuning.

While working on this, I added support to the basic trainer for
constructing models on the meta device, materializing them on the actual
device, and loading the state dictionaries directly to the tensors on
the device.

The result in that loading checkpoint with this trainer is much faster.
The pipeline trainer already does this.