Reorganized Dataset Templates
The dataset templates have been updated to use a more consistent naming
scheme and have generally been cleaned up and tested.

A few new dataset definitions have been added and "abridged" version
added to the giant dataset, as to provide a faster loading option for
shorter experiments.

While working on this, the "dataset" command in the CLI has been refined
a bit more. It's still a bit rough, but better.

I'm also include a fix for the llama tutorial, where the change in
module organization missed a reference to the new location for
callbacks.