Dataset re-org, stage 2
Datasets are now first-class Forgather projects, rather than templates
to be inlined in another project.

Previously, the template for including in another project and the actual
project were in completely seperate locations, making maintainance more
awkward than need be.

This makes the interface bettween datasets and their consumers more
distinct, which should help with maintainance.

The primary reason for this change is that it facilitates reusing the
dataset definitions outside of a strict training project, as I have
something specific in mind for this.

As usual, this required moving a great many files around, but in the
end, I believe it's worth it.

While working on this, some maintinance was done on the main Forgather
libraries to make things a little cleaner -- and added the ability to
use inter-project dependency injections. Datasets require an injected
tokenizer object and tokenizer args.

Much imporoved documentation has been written for working with example
datasets as well.