Added dynamic arguments
With dynamic arguments, you can define a set of arguments avaiable to
the CLI in a configuration and the values specified on the CLI will be
injected into the template. For example, by adding a "--max-steps"
argument, you can provide a knob for a quick test, without having to
modify the templates. In theory, this should also make a hyper-parameter
search easier, as it does not require creating a config template for
each setting.

Added a few examples of usage of dynamic arguments to the example
projects.

Add !dlist type (Dictionary List). This is a list, which is defined as a
dictionary. Why do such a thing? This makes is much easier to override
values in a list with the templating system -- or delete them from a
list, by name.

Updated systax reference for new !dlist, with examples.