Harmonize gradient_checkpointing option with HF
HF PretrainedModel uses the training argument "gradient_checkpointing"
to enable activation checkpointing, which triggers a call to to
PretrainedModel.gradient_checkpointing_enable(). This in turn checks to
see if the model support checkpointing, via the
supports_gradient_checkpointing class attribute. If true, it iterates
over all modules, setting "gradient_checkpointing" to True, if this
attribute exists. The module is then expected to call the checkpoint
funciton only when this is set, wherre the checkpoint function is also
set on the module.

We implement this convention now, where checkpoint_layer_stack uses the
new value set by gradient_checkpointing_enable(), otherwise it uses
whatever the default was set to.