Reworked base trainer templates
This started with adding universal support for the --max-steps and
--save-strategy dynamic args. The intent is to be able to automate
testing of all of the projects and their configuration, providing the
means to run short tests, without actually saving anything to disk.

This required reworking the base trainer and training script templates
and this turned into a general maintenance sweep of the trainer
definitions and training script mechanics.

- Added universal training script support for --max-steps and --save-strategy
- Removed trainer's namespace and moved arguments into "ns" namspace.
- Removed default implementation of create_new_model block in training
script. Now children are always responsible for specifying how to do this.
- Loading model weights is now exclusively through the checkpoint
loader. Previously, there was a fair amount of use of using the HF API.
- Cleaned up many of the templates, removed unused cruft and making
things more consistent.