Created IterableDataset length wrapper
When a Huggingface map-style dataset is converted to an interable
dataset and mapped, it loses it's __len__ attribute. This is fair enough, when the
map function could potentially return output which are not 1-to-1, but
if we know that the outputs are 1-1, it simplifies things if we can get
the length of the dataset. For example, determining the length of an
epoch and showing progress.

A wrapper has been created for this use-case, where __len__ returns the
length of the original dataset -- and properly forwards calls and
attributes.