Call distributed.destroy_process_group() on script exit
We get warnings for not calling destroy_process_group() on each node
before exit. This had previously been enabled, but always resulted in a
segfault. The warning seemed preferable.

I tried enabling this code again and -- no segfaults!?

So, we now call this before exit. If the segfault is seen again, we can
undo this change.