Reorganize templatelib into a more coherent namespace
The consistency of the tempaltelib namespace had much to be desired. I
have been putting this off, as this type of cleanup involves touching
many file, with a high risk of breaking things, but better do it sooner,
rather than later, as this type of problem grows with time.

I think I have everything working and the namespace is as least more
consistent now.

I renamed "type.yaml" to be more specific, "config_type.yaml," as it's
really the base class for configurations, as well as adding comments to
this file and it's children, clarifying their role.

While looking for broken templates, I found a couple I had missed in a
previous sweep, and fixed them.

It would not surprise me if something subtle has broken, but at least
everything that is expected to parse does now.