Add support for logging gradient-norm
Add support for logging gradient-norm, even in the case where we are
fusingthe backward and optimizer steps, which can't use the computed
value for grad-norm-clipping.

While working on this change, a fair amount of refactoring of the main
training loop took place. The logic for periodic logging, evaluating,
and saving has been moved to a new function, _mayby_log_save_evaluate().

Much of the callback control logic from the main loop has been greatly
simplified.

We have removed the code for adding evaluation metrics to the log. These
don't belong there, as the callbacks don't expect these to be in a
unified log. on_log_step() expects the train logs to be in "log" and
on_eval_step() expects the eval metrics in "metrics." This is more
consistent with how the Huggingface Trainer implementation handles this.

The periodic function code has largely been rewritten to better fit with
the refactored log-save-eval function, where these operations can be
triggered by either the periodic-functions or from control messages from
the callbacks.

All of the logic for computing the metrics to log has been moved to
_log_step().

_train_step() is now expected to return both loss and total_norm, for
which both should have been processed by any distributed collective
operations, where applicable.

When we are fusing the optimizer and backward step, we compute
accumulate the squared-sums of each tensor in the optimizer hook, which
is implemented as a partial function. This allows us to compute the
total norm, while only having a single materialized gradient at a time.
Once we finish the backward step, we apply sqrt() to the total in
_clip_grad_norm(), where we return the value, but don't perform clipping
-- because we have already performed the optimizer step.

In the main training loop, we keep track of both the gradient norms and
the loss values in lists, where previously we used a scalar for the
loss. This format allows for other useful transforms on these values.
For example, we report both the RMS gradient norm, for each step between
logs steps, and the max gradient norm for that interval.

tb_logger has been updated to support the new values and has been made
more flexible in general. Rather than having a hard-coded set of scalars
to log, it takes a list describing which metrics to report and how to
report them. This allows for generating new metrics from other values,
for example, perplexity is computed via a hook using "loss."

Default callbacks has also been updated to report the RMS grad-norm.
It's not nearly as flexible as the Tensor Board logger. Perhaps
something to improve upon another day.