Added experimental support for torchdata.node
Added basic support for the torchdata.node API
https://docs.pytorch.org/data/beta/torchdata.nodes.html

Right now, just pass a node (or a DataLoader) instead of a dataset and
it will be used as-is, without wrapping it in a Dataloader. This
obviously requires that this object defines how to produce the final
data examples to pass to the model.