Added very basic "model" command to the CLI
Right now, it can construct a model on a target device and print it as
well as running a test forward and backward pass.

Later, I'll add profiling functions and make it more flexible, as causal
LMs, with tokenizers, are not the only kind of model...

This at least helps with basic diagnostics, when creating a new model
type.