Workaround for Samantah dataset
Huggingface recently changes the datasets library to disallow the use of
"trust_remote_code." As the Samatha dataset is designed to use a custom
script for converting the json files to a dataset, this has broken our
example code using the Samantha dataset.

This is a quick and drity work-around, where we load the json files
directly from the repo and build them into a dataset via custom code in
samantha.yaml.

It works, but it could be better. It should probably be implemented as a
DatasetBuilder -- and it would be nice to add the ability to mere the
extra examples from "data" into the dataset.

I'll take a look at that later. This at least allows using the dataset
for now.