Major refactoring of training checkpointing
The checkpointing functionality has been moved into a CheckpointManager
class, with the end goal of being able to dynamically swap out
checkpointing implementations, as I would like to add support for
torch.distributed.checkpoint.

This brings us one step closer. It's mostly the same code from the
trainer, but greatly cleaned up and simplified.

The AccelTrainer functionality /mostly/ works right now. It seems that
they have disabled StatefulDataloaders by default. A quick and dirty
hack shows that it does work, but this approach seems fragile. I'll see
about a proper fix shortly.

The checkpoint unit-test has been updated to have tests for saving and
resuming stateful dataloaders, as this was missing.

While working on this, I have also made some progress with general code
cleanup in the trainers, with a fair amount of improvement with typing
issues in the static analyzer.