Work around for Accelerate bug
It seems that setting use_stateful_dataloader=True and num_workers > 0
causes Accelerate to bork. See:

https://github.com/huggingface/accelerate/issues/3110

This is what I ran into last night, so it's not my bug. The solution?
Disable concurrent dataloading (num_workers = 0), if you want to be able
to restore your Dataloader state when checkpointing. This bug has been
around for awile, so I'm surprised that it has not had more traction.

While working on this, I confirmed that we can call
torch.distributed.init_process_group() before constructing an
Accelerator instance, without causing it to barf. So, I have moved the
Accelerator setup from the distributed-environment wrapper to the
acclerate trainer definition, where it belongs.

And boo to not being able to combine stateful dataloaders and
num_workers > 0. Enough said.