Minimal working Pipeline grad accumulation
Implemented gradient accumulation in pipeline trainer.

It seems that the
https://github.com/pytorch/pytorch/blob/main/torch/distributed/pipelining/stage.py#L567
(scale_grads()) automatically scales grads by the number of microbatches
by dividing the gradients by N. This is completely broken when using
gradient accumulation, as this causes the total accumulated gradient to
be rescaled by this factor at each step, resulting in the total
gradient, at the end of gradient accumulation, being too small.

There is a scheduler option, "scale_grads," which defaults to True,
which controls this. As we already need to wrap the loss function to
scale the gradients by the number of accumulation steps, we can resolve
this by disabling "scale_grads" and including the number of microbatches
in the total scale factor.

When compared to the equivalent test run, without gradient accumulation,
but not Pipeline Parallel, this prodcues results much closer to what is
expected.

Skimming over the Torch Titan code, it looks like they probably have
this bug, where gradient accumulation is not computed correctly. I only
noticed, as I have a reference run to compare against.

I'll try to confirm if this bug is present and offer a PR, if found to
be the case.