Further refactoring of DistributedEnvironment
We add a function for getting a viable barrier function for the current
device.

We add a static dataclass, which just provides info, without
initializing the envrionment.

We move the code for initializing distributed envrionment parameters
from envrionment variables into a seperate function.

We add a "from_env()" function, which returns a static distributed
envrionment, read from env variables.