Fix issue with preprocessor ordering
As always, it just can't be simple. Performing the toml block transform
before the regular preprocssing can cause some of the block statements
to become comments -- or embedded in line statements.

This is resolved by performing line-statement preprocessing first.

The tiny stories project was still using the '-' formatting flag, which
has been removed. I updated and fixed the config.