Maintaince on tutorial
Converted more components to new syntax.

To things easier to comprehed, for first exposure to Forgather, I have
replaced the default config, using the dynamic llama definition, to
import the equivalent model from the example models project. This
reduces the amount of things someone will need to process to understand
what's going on.

The experimental Llama still uses the original model def, as it's
purpose is to show how to extend a model, directly from a trainer
config.

I noted that the HF Llama config is broken. Added to TODO list to fix.