Added linked checkpoint weight function
For the Huggingface .from_pretrained() API to work, the model's weights
need to be located in the root of the model directory, along with
configs, tokenizers, model code, etc.

When training, we save checkpoints to a "checkpoints" sub-directory,
which .from_pretrained() can't find. To address this, one can
move/copy/symlink the checkpoint weights into the model root directory,
but doing this manually is a hassle.

Solution: Add a function to located the latest checkpoint weights and
make symlinks in the model's root directory, pointing to the latest
checkpoint weights.

Obviously, we don't want to replace real weights in the root directory,
if they exist, so a safe default, don't clobber the existing weights, has
been added, with an option to override.

The Tiny Llama tutorial had not been using checkpointing, but rather
saved the weights separately at the end of training, to address this in
the tutorial. Specifically, we want to demo that you can use
.from_pretrained() to load your newly trained model. This is the only
example that does this, and it's not the recommended practice. Now, we
can provide an example that works like all of the others AND can load
the model with .from_pretrined() at the end.

A more common use-case is to test model checkpoints with inference tools,
while training is ongoing. This is normally something of a pain, but it
need not be!

The next step would be to add a command to "forgather" to perform this
action from the CLI. But not in this commit.