Fix name error in warning
If the configured sequence length provided to the data-collator exceeds
the maximum length, reproted by the tokenizer, the code is configured to
emmit a warning. We dump both values in the warning and once of the
values was missing "self," raising an exception.

It's an open question if raising an exception is not the better
option. Models with fixed length positional embeddings are likely to
raise an exception anyway and it's likely that the warning will be
lost in the noise from the exception, which does not clearly point to
the reason for the failure.

Some encoders, like ALiBi and NOPE, will still work.

Switching to raising an explicit exception, rathe than just fixing the
bug. One can always update the tokenizer definition, if one wants to
experimetn.