Fixed a bug in the YAML encoder
Recently, we added the ability to define a list, without any arguments,
in a configuration, which makes it easy to define an empty list, which
can be extended in a sub-template.

The YAML encoder assumed that there is always a list argument present,
so raises an exception, if you have such an empty list.

This has been updated to handle the case where the list argument is
implicit.

While working on this, I noticed that we don't have the same
functionality for !tuple, so this now can take an empty sequence as
well.