New project for comparing titan palallelisms
This adds a new project for comparing various parallel modes, where the
mode is the only variable.

While testing, I confirmed that the bug I found in my own pipeline
trainer, with respect to gradient accumulation, does exist in torch
titan.

Details are in the README.md

I'll submit a bug and a PR momentarily for fixing the issue of the
torchtitan side.

This change also includes a new feature for specifying envrionment
variables to inject into a training job in the configuration and a
(disabled) fix for my version of the titan trainer. There's a flag to
enable it, but that not much help, without fixing torch titan as well.