Use native Torch Titan Trainer class
For the initial bring-up, allowing one to import custom optimizers,
lr-schedulers, and datasets, I created a custom sub-class of the
torchtitan Trainer class. This works, but it did require duplicating
much of the constructor code -- something of a moving-target, WRT
maintainance.

To address this, we now build and register a TrainSpec before
constructing a native torchtitan Trainer instance, which achieves the
same end, but without having to duplicate the constructor.

This is still less than perfect. A number of "helper" functions have
been added to assist with the various "build" functions, expected by
TrainSpec.

The titan Validator is poorly abstracted WRT building the validation
dataloader. It just reads the job_config and builds it from that, which
does not make it easy to inject a custom dataset. To address this, I
created a sub-class and modifed the constructor to allow one to inject
the dataset. This will probably be a maintainance headache, but at least
the code duplication is minimal.

There was a similar issue for the titan lr-scheduler container, which
only supports LambdaLR lr-schedulers. To support everything else, I
duplicated the original implementation, modifying it to take a
lr-scheduler factory argument.

Finally, as Titan is something of a moving target anyway, I sync'd the
latest from mainline and fixed the latest breakages.

I'm sure things will still break, but this at least has less duplicated
logic to maintain.