Updated distributed.py to handle multi-node config
The existing distributed.py supports multi-worker on a single node, but
had issues dealing with the multi-node setting.

main_process_first() is intended to only run on LOCAL_RANK == 0 on each
node, but the barrier was waiting for all ranks, not just local ones.
This has been resolved by adding a function to create a local process
group for each node and updating barrier_fn() to take a process group as
an argument, when constructing a barrier function.

A potential remaining issue is that we assume symmetry among nodes, WRT
the number of workers per node. If this assumption is violated, the
present implementation will not work correctly.

TODO: Fix above issue.