Make missing checkpoint files less verbose
When loading a checkpoint, where we only have the model weights, the
existing code dumps a stack trace for each missing file.

This is actually a pretty common scenario, when starting to train on a
pretrained model. Reduce the noise by commenting out the code to dump
the stack trace.