Train at 16K context length!
It seems that the H.P. Lovecraft example can be pushed to train a 7B
model with a context length of up to 16K on a single GPU!? That's full
fine-tuning, bfloat16, on a 24 GB GPU.

I have reorganized the tutorial a bit and documented hints for doing
this in the Extra Credit section. I'll see about creating an example
project with this setup, so you don't have to create it from scratch.

First, some experimentation is in order...