Add new PAD token to converted models
When converting Llama to Fg format, some models do not have a defined
PAD token. To simplify handling, we just assigned the value of the EOS
token, but this can cause issues for the model learning to know when to
stop.

Instead, add a new zero'd token embedding for the missing PAD token, if
required.

I also noted that some chat templates (Mistral Instruct) expects
bos_token and eos_token to be injected into the Jinja2 template.

Modified the server, as well as our Samantha and Orca examples to inject
these values from the tokenizer.

As it may be useful to train a model with EOS tacked on after the
assistant's <|im_end|>, I added an alternative version of this template,
uses the new "eos_token" var to add EOS.